+++ jgclark.QuickCapture/src/quickCaptureHelpers.js
@@ -2,16 +2,15 @@
 // ----------------------------------------------------------------------------
 // Helpers for QuickCapture plugin for NotePlan
 // by Jonathan Clark
-// last update 2025-07-20 for v0.16.2 by @jgclark
+// last update 2025-07-28 for v0.17.0 by @jgclark
 // ----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
-import { getDateOptions, getFilenameDateStrFromDisplayDateStr, isValidCalendarNoteTitleStr } from '@helpers/dateTime'
-import { getRelativeDates, getWeekOptions } from '@helpers/NPdateTime'
+import { isValidCalendarNoteTitleStr } from '@helpers/dateTime'
+import { getRelativeDates } from '@helpers/NPdateTime'
 import { clo, logInfo, logDebug, logError, logTimer, logWarn } from '@helpers/dev'
-import { allNotesSortedByChanged, calendarNotesSortedByChanged } from '@helpers/note'
-import { openNoteByFilename } from '@helpers/NPnote'
-import { displayTitleWithRelDate, showMessage } from '@helpers/userInput'
+import { getOrMakeCalendarNote } from '@helpers/NPnote'
+import { chooseNoteV2, displayTitleWithRelDate, showMessage } from '@helpers/userInput'
 
 //----------------------------------------------------------------------------
 // helpers
@@ -78,17 +77,19 @@ export async function getQuickCaptureSettings(useDefaultsIfNecessary: boolean = 
  * If a desired Calendar note doesn't already exist this now attempts to create it first.
  * Note: Send param 'allNotesIn' if the generation of that list can be more efficiently done before now. Otherwise it will generated a sorted list of all notes.
  * Note: There's deliberately no try/catch so that failure can stop processing.
- * TODO(Later): Hopefully @EM will allow future calendar notes to be created, and then some of this handling won't be needed.
  * TODO: Move this to helpers/NPNote.
  * See https://discord.com/channels/763107030223290449/1243973539296579686
+ * TODO(Later): Hopefully @EM will allow future calendar notes to be created, and then some of this handling won't be needed.
  * @param {string} purpose to show to user in dialog title 'Select note for new X'
- * @param {string?} noteTitleArg to match against note titles
- * @param {boolean?} justCalendarNotes? (default: false)
- * @param {boolean?} addFutureCalendarNotes? (default: true)
+ * @param {string?} noteTitleArg to match against note titles. If not given, will ask user to select from all note titles (excluding the Trash).
  * @param {Array<TNote>?} notesIn
  * @returns {TNote} note
  */
-export async function getNoteFromParamOrUser(purpose: string, noteTitleArg?: string = '', justCalendarNotes: boolean = false, addFutureCalendarNotes: boolean = true, notesIn?: Array<TNote>): Promise<TNote | null> {
+export async function getNoteFromParamOrUser(
+  purpose: string,
+  noteTitleArg: string = '',
+  notesIn?: Array<TNote>,
+): Promise<TNote | null> {
   // Note: deliberately no try/catch so that failure can stop processing
   const startTime = new Date()
   let note: TNote | null
@@ -106,95 +107,45 @@ export async function getNoteFromParamOrUser(purpose: string, noteTitleArg?: str
         break
       }
     }
-
     // If this has already found a note, return it
     if (note) {
       logDebug('getNoteFromParamOrUser', `- Found note from noteTitleArg '${noteTitleArg}'`)
       return note
     }
 
-    // Otherwise check to see if it is of the *form* of a Calendar note string
+    // Now check to see if the noteTitleArg is of the *form* of a Calendar note string
     if (isValidCalendarNoteTitleStr(noteTitleArg)) {
       noteTitleArgIsCalendarNote = true
-      logDebug('getNoteFromParamOrUser', `- Note '${noteTitleArg}'is of the form of a Calendar note string. Will attempt to create it.`)
-      const wantedFilename = `${getFilenameDateStrFromDisplayDateStr(noteTitleArg)}.${DataStore.defaultFileExtension}`
-      // $FlowIgnore[incompatible-type] straight away test for null return
-      note = await openNoteByFilename(wantedFilename, {})
+      logDebug('getNoteFromParamOrUser', `- Note is of the form of a Calendar note string. Will attempt to create it.`)
 
-      if (note) {
-        logDebug('getNoteFromParamOrUser', `- Made new note with filename ${note.filename}`)
-      } else {
-        logWarn('getNoteFromParamOrUser', `Couldn't find Calendar note with title '${noteTitleArg}'. Will suggest a work around to user.`)
+      // Test to see if we can get this calendar note
+      // $FlowIgnore[incompatible-type] straight away test for null return
+      note = getOrMakeCalendarNote(noteTitleArg)
+      if (!note) {
+        logWarn('getNoteFromParamOrUser', `Couldn't find or make Calendar note with title '${noteTitleArg}'. Will suggest a work around to user.`)
         throw new Error(
-          `I can't find Calendar note '${noteTitleArg}', and unfortunately I can't create it for you.\nPlease create it by navigating to it, and adding any content, and then re-run this command.`,
+          `I can't find Calendar note '${noteTitleArg}', and unfortunately I have tried and failed to create it for you.\nPlease create it by navigating to it, and adding any content, and then re-run this command.`,
         )
       }
     }
-  }
 
-  // Now try to find one
-  logDebug('getNoteFromParamOrUser', `- Couldn't find note with title '${noteTitleArg}'.`)
+    // Now try to find wanted regular note
+    // logDebug('getNoteFromParamOrUser', `- Couldn't find note with title '${noteTitleArg}'.`)
 
-  // If wanted, get dates of future daily notes for the next month, and future weekly notes for the next 6 months
-  // TODO: decide whether to keep this suggestion from DBW:
-  const futureDateStrings = addFutureCalendarNotes ? [...getDateOptions(), ...getWeekOptions()] : []
+    // Preferably we'll use the last parameter, but if not calculate the list of notes to check
+    const notesToCheck = getNotesToCheck(notesIn)
 
-  // First check to see if it is of the *form* of a Calendar note string
-  if (isValidCalendarNoteTitleStr(noteTitleArg)) {
-    noteTitleArgIsCalendarNote = true
-    logDebug('getNoteFromParamOrUser', `- Note is of the form of a Calendar note string. Will attempt to create it.`)
-    const wantedFilename = `${getFilenameDateStrFromDisplayDateStr(noteTitleArg)}.${DataStore.defaultFileExtension}`
-    // $FlowIgnore[incompatible-type] straight away test for null return
-    note = await openNoteByFilename(wantedFilename, {})
+    const matchingNotes = notesToCheck.filter((n) => n.title?.toLowerCase() === noteTitleArg.toLowerCase())
+    logDebug('getNoteFromParamOrUser', `Found ${matchingNotes.length} matching notes with title '${noteTitleArg}'. Will use most recently changed note.`)
+    note = matchingNotes[0]
 
-    if (note) {
-      logDebug('getNoteFromParamOrUser', `- Made new note with filename ${note.filename}`)
-    } else {
-      logWarn('getNoteFromParamOrUser', `Couldn't find Calendar note with title '${noteTitleArg}'. Will suggest a work around to user.`)
-      throw new Error(
-        `I can't find Calendar note '${noteTitleArg}', and unfortunately I can't create it for you.\nPlease create it by navigating to it, and adding any content, and then re-run this command.`,
-      )
-    }
   } else {
-    // If not, form list of notes to check against / offer
-    let allNotesToUse: Array<TNote> = []
-    if (notesIn) {
-      allNotesToUse = notesIn
-      logTimer('getNoteFromParamOrUser', startTime, `- Used 4th param which has ${notesIn.length} entries`)
-    } else {
-      allNotesToUse = justCalendarNotes ? calendarNotesSortedByChanged() : allNotesSortedByChanged()
-      logTimer('getNoteFromParamOrUser', startTime, `- Got large note array of all ${justCalendarNotes ? 'calendar' : ''} notes`)
-    }
-
-    // Preferably use 4th parameter, but if not calculate the list
-    if (!note && allNotesToUse) {
-      // Note: Because of NP architecture, it's possible to have several notes with the same title; the first match is used.
-      const noteTitleToMatch = noteTitleArg
-      logDebug('getNoteFromParamOrUser', `- noteTitleToMatch = ${noteTitleToMatch}`)
-      // Change YYYY-MM-DD to YYYYMMDD format if needed.
-      const wantedNotes = allNotesToUse.filter((n) => displayTitleWithRelDate(n, false) === noteTitleToMatch)
-      if (wantedNotes.length > 1) logInfo('getNoteFromParamOrUser', `Found ${wantedNotes.length} matching notes with title '${noteTitleArg}'. Will use most recently changed note.`)
-    }
-
-    logTimer('getNoteFromParamOrUser', startTime, `- mid point`)
-
-    if (!note) {
-      // Couldn't find the note. 
-      // If this looks to be a Calendar note then there's a bad work around for this.
-      if (noteTitleArgIsCalendarNote) {
-        logWarn('getNoteFromParamOrUser', `Couldn't find Calendar note with title '${noteTitleArg}'. Will suggest a work around to user.`)
-        throw new Error(`I can't find Calendar note '${noteTitleArg}', and unfortunately I can't create it for you.\nPlease create it by navigating to it, and adding any content, and then re-run this command.`)
-      }
-
-      if (noteTitleArg !== '') {
-        logDebug('getNoteFromParamOrUser', `Couldn't find regular note with title '${noteTitleArg}'. Will prompt user instead.`)
-      }
-
-      const notesList = allNotesToUse.map((n) => displayTitleWithRelDate(n)).filter(Boolean)
-      const result = await CommandBar.showOptions(notesList, `Select note for new ${purpose}`)
-      if (typeof result !== 'boolean') {
-        note = allNotesToUse[result.index]
-      }
+    // We need to ask user to select from all notes
+    // Preferably we'll use the last parameter, but if not calculate the list of notes to check
+    const notesToCheck = getNotesToCheck(notesIn)
+    const result = await chooseNoteV2(`Select note for new ${purpose}`, notesIn, true, true, false, false)
+    if (typeof result === 'boolean') {
+      note = notesToCheck[result.index]
     }
   }
   // Double-check this is a valid note
@@ -204,4 +155,11 @@ export async function getNoteFromParamOrUser(purpose: string, noteTitleArg?: str
 
   logTimer('getNoteFromParamOrUser', startTime, `-> note '${displayTitleWithRelDate(note)}'`)
   return note
+}
+
+function getNotesToCheck(notesIn?: Array<TNote>): Array<TNote> {
+  if (notesIn) {
+    return notesIn
+  }
+  return DataStore.projectNotes.filter((n) => !n.filename.startsWith('@Trash'))
 }